# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "High-performance AI gateway routing requests to multiple LLM providers (OpenAI, Anthropic, Gemini, Groq, OpenRouter, DeepSeek, Z.ai, xAI, MiniMax, Oracle, Ollama). Drop-in OpenAI-compatible API.",
    "title": "GoModel API",
    "contact": {},
    "version": "1.0"
  },
  "paths": {
...
    "/admin/api/v1/cache/overview": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get cached-only usage overview",
        "parameters": [
          {
            "description": "Number of days (default 30)",
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Start date (YYYY-MM-DD)",
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Grouping interval: daily, weekly, monthly, yearly (default daily)",
            "name": "interval",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by tracked user path subtree",
            "name": "user_path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Cache mode filter: uncached, cached, all (cache overview always uses cached mode)",
            "name": "cache_mode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/usage.CacheOverview"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/admin/api/v1/usage/daily": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get usage breakdown by period",
        "parameters": [
          {
            "description": "Number of days (default 30)",
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Start date (YYYY-MM-DD)",
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Grouping interval: daily, weekly, monthly, yearly (default daily)",
            "name": "interval",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by tracked user path subtree",
            "name": "user_path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Cache mode filter: uncached, cached, all (default uncached)",
            "name": "cache_mode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/usage.DailyUsage"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/admin/api/v1/usage/log": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get paginated usage log entries",
        "parameters": [
          {
            "description": "Number of days (default 30)",
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Start date (YYYY-MM-DD)",
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by model name",
            "name": "model",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by provider name or provider type",
            "name": "provider",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by tracked user path subtree",
            "name": "user_path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Cache mode filter: uncached, cached, all (default uncached)",
            "name": "cache_mode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Search across model, provider, request_id, provider_id",
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Page size (default 50, max 200)",
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Offset for pagination",
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/usage.UsageLogResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/admin/api/v1/usage/models": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get usage breakdown by model",
        "parameters": [
          {
            "description": "Number of days (default 30)",
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Start date (YYYY-MM-DD)",
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by tracked user path subtree",
            "name": "user_path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Cache mode filter: uncached, cached, all (default uncached)",
            "name": "cache_mode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/usage.ModelUsage"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/admin/api/v1/usage/recalculate-pricing": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Recalculate stored usage costs from current model pricing metadata",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/admin.recalculatePricingRequest"
              }
            }
          },
          "description": "Recalculation filters and confirmation",
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/usage.RecalculatePricingResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/admin/api/v1/usage/summary": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get usage summary",
        "parameters": [
          {
            "description": "Number of days (default 30)",
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Start date (YYYY-MM-DD)",
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by tracked user path subtree",
            "name": "user_path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Cache mode filter: uncached, cached, all (default uncached)",
            "name": "cache_mode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/usage.UsageSummary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/admin/api/v1/usage/user-paths": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get usage breakdown by user path",
        "parameters": [
          {
            "description": "Number of days (default 30)",
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Start date (YYYY-MM-DD)",
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by tracked user path subtree",
            "name": "user_path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Cache mode filter: uncached, cached, all (default uncached)",
            "name": "cache_mode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/usage.UserPathUsage"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/core.GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://gomodel.example.com",
      "description": "GoModel HTTPS deployment"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local GoModel"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "core.GatewayError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string"
          },
          "param": {
            "type": "string",
            "nullable": true
          },
          "provider": {
            "type": "string"
          },
          "status_code": {
            "type": "integer"
          },
          "type": {
            "$ref": "#/components/schemas/core.ErrorType"
          }
        }
      },
...
      "usage.CacheOverview": {
        "type": "object",
        "properties": {
          "daily": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/usage.CacheOverviewDaily"
            }
          },
          "summary": {
            "$ref": "#/components/schemas/usage.CacheOverviewSummary"
          }
        }
      },
...
      "usage.DailyUsage": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "input_cost": {
            "type": "number"
          },
          "input_tokens": {
            "type": "integer"
          },
          "output_cost": {
            "type": "number"
          },
          "output_tokens": {
            "type": "integer"
          },
          "requests": {
            "type": "integer"
          },
          "total_cost": {
            "type": "number"
          },
          "total_tokens": {
            "type": "integer"
          }
        }
      },
      "usage.ModelUsage": {
        "type": "object",
        "properties": {
          "input_cost": {
            "type": "number"
          },
          "input_tokens": {
            "type": "integer"
          },
          "model": {
            "type": "string"
          },
          "output_cost": {
            "type": "number"
          },
          "output_tokens": {
            "type": "integer"
          },
          "provider": {
            "type": "string"
          },
          "provider_name": {
            "type": "string"
          },
          "total_cost": {
            "type": "number"
          }
        }
      },
...
      "usage.UsageLogResult": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/usage.UsageLogEntry"
            }
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}